Arreglar backend/Agregar login al frontend
He agregado los primeros dtos, he agregado un config para la encriptacion de contrase;as, modifique un atributo llamado pago de double a bigdecimal, agregué la relacion maytomany al model usuario con la tabla usuario_rol, agregue un auth controller y service para lo del login, revisen el applicationproperties el mio tiene contrase;a diferente. Sobre el frontend, ya hice el login funcional para ello debe crear un usuario haganlo desde el postman, pueden avanzar con el registro

+++ backend/sistema-citas-dentales/src/main/java/com/dental/backend/model/Pago.java
@@ -4,11 +4,11 @@ import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 import com.dental.backend.model.enums.MetodoPago;
-
+import java.math.BigDecimal;
 @Entity
 @Table(name = "pagos")
 public class Pago {
-
+	
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_pago")
@@ -19,7 +19,7 @@ public class Pago {
     private Cita cita;
 
     @Column(precision = 10, scale = 2)
-    private Double monto;
+    private BigDecimal monto;
 
     @Enumerated(EnumType.STRING)
     private MetodoPago metodo;
@@ -51,13 +51,14 @@ public class Pago {
 		this.cita = cita;
 	}
 
-	public Double getMonto() {
-		return monto;
-	}
+	public BigDecimal getMonto() {
+        return monto;
+    }
 
-	public void setMonto(Double monto) {
-		this.monto = monto;
-	}
+    // 3. CAMBIO EN EL SETTER
+    public void setMonto(BigDecimal monto) {
+        this.monto = monto;
+    }
 
 	public MetodoPago getMetodo() {
 		return metodo;
